+++ NEWS.md
@@ -1,3 +1,6 @@
+* Added functions to calculate indifference curves for preferences graph.
+* Added functions to calculate iso lines on energy, cost, and preferences graphs.
+* Added functions to calculate paths for rebound analysis on energy, cost, and preferences graphs.
 * Added functions to calculate all rebound parameters.
 * Added `sample_eeu_data_path()` function to give the path to an Excel file
   containing same energy efficiency upgrade data.